issue-997: extract voice contract/runtime
