+++ Cargo.toml
@@ -25,6 +25,7 @@ members = [
     "crates/tau-dashboard",
     "crates/tau-custom-command",
     "crates/tau-browser-automation",
+    "crates/tau-voice",
     "crates/tau-memory",
   "crates/tau-coding-agent",
 ]
