+++ crates/tau-coding-agent/Cargo.toml
@@ -38,6 +38,7 @@ tau-github-issues = { path = "../tau-github-issues" }
 tau-dashboard = { path = "../tau-dashboard" }
 tau-custom-command = { path = "../tau-custom-command" }
 tau-browser-automation = { path = "../tau-browser-automation" }
+tau-voice = { path = "../tau-voice" }
 tau-agent-core = { path = "../tau-agent-core" }
 tau-ai = { path = "../tau-ai" }
 reqwest.workspace = true
