+++ crates/tau-voice/src/lib.rs
@@ -1,0 +1,2 @@
+pub mod voice_contract;
+pub mod voice_runtime;
